+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Updated translations ([#127](https://codeberg.org/silverpill/mitra-web/pulls/127)).
 - Require NodeJS 20.
+- Disable "Vote" button while vote is being processed.
 
 ## [4.5.1] - 2025-06-23