+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Read minimum payment amount from `/api/v1/instance` data.
 - Don't display quoted post if author is muted.
 - Don't convert instance title to uppercase on landing page.
+- Changed icon for Local timeline.
 
 ## [4.5.1] - 2025-06-23