+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added Persian translation ([#127](https://codeberg.org/silverpill/mitra-web/pulls/127)).
+- Support uploading from URL.
 
 ### Changed