+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.6.0] - 2025-07-16
+
 ### Added
 
 - Added Persian translation ([#127](https://codeberg.org/silverpill/mitra-web/pulls/127)).